+++ src/components/UserDetail/UserDetail.js
@@ -1,7 +1,11 @@
-import _ from 'lodash';
+import { get } from 'lodash';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Link } from 'react-router-dom';
+import {
+  FormattedDate,
+  FormattedMessage,
+} from 'react-intl';
 
 import {
   Col,
@@ -9,8 +13,8 @@ import {
   Row,
 } from '@folio/stripes/components';
 
-import { FormattedDate, FormattedMessage } from 'react-intl';
 import { getFullName } from '../../util';
+
 import css from './UserDetail.css';
 
 class UserDetail extends React.Component {
@@ -29,6 +33,10 @@ class UserDetail extends React.Component {
       records: 'accounts',
       path: 'accounts?query=(userId=!{user.id} and status.name<>Closed)&limit=100',
     },
+    openRequests: {
+      type: 'okapi',
+      path: 'circulation/requests?query=(requesterId=!{user.id} and status=Open)&limit=100',
+    },
   });
 
   static propTypes = {
@@ -45,8 +53,14 @@ class UserDetail extends React.Component {
       openAccounts: PropTypes.shape({
         records: PropTypes.arrayOf(PropTypes.object),
       }),
+      openRequests: PropTypes.shape({
+        records: PropTypes.arrayOf(PropTypes.object),
+      }),
     }),
     renderLoans: PropTypes.bool,
+    stripes: PropTypes.shape({
+      hasPerm: PropTypes.func.isRequired,
+    }).isRequired,
   };
 
   getUserValue = (user) => {
@@ -62,14 +76,15 @@ class UserDetail extends React.Component {
             {getFullName(user)}
           </strong>
         </Link>
-        <strong>
-          <FormattedMessage id="ui-checkout.user.detail.barcode" />
-        </strong>
+        <FormattedMessage
+          id="ui-checkout.user.detail.barcode"
+          tagName="strong"
+        />
         {' '}
         {user.barcode ? (<Link to={path}>{user.barcode}</Link>) : '-'}
       </span>
     );
-  }
+  };
 
   renderLoans() {
     const {
@@ -79,12 +94,12 @@ class UserDetail extends React.Component {
     } = this.props;
 
     if (!renderLoans) return null;
-    const openLoansCount = _.get(resources.openLoansCount, ['records', '0', 'totalRecords'], 0);
+
+    const openLoansCount = get(resources.openLoansCount, ['records', '0', 'totalRecords'], 0);
     const openLoansPath = `/users/view/${user.id}?layer=open-loans&query=`;
     const openLoansLink = <Link to={openLoansPath}>{openLoansCount}</Link>;
-
-    const patronGroups = _.get(resources, ['patronGroups', 'records', 0, 'group'], '');
-    const openAccounts = _.get(resources, ['openAccounts', 'records'], []);
+    const patronGroups = get(resources, ['patronGroups', 'records', 0, 'group'], '');
+    const openAccounts = get(resources, ['openAccounts', 'records'], []);
     const openAccountsPath = `/users/view/${user.id}?layer=open-accounts&filters=pg.${patronGroups}`;
     const owedAmount = openAccounts.reduce((owed, { remaining }) => {
       return owed + parseFloat(remaining);
@@ -109,8 +124,45 @@ class UserDetail extends React.Component {
               value={openAccountsLink}
             />
           </Col>
+          <Col xs={4}>
+            <KeyValue
+              label={<FormattedMessage id="ui-checkout.openRequests" />}
+              value={this.renderOpenRequests()}
+            />
+          </Col>
         </Row>
       </div>
+    );
+  }
+
+  renderOpenRequests() {
+    const {
+      resources,
+      stripes,
+      user,
+    } = this.props;
+
+    const openRequestsCount = get(resources.openRequests, ['records', '0', 'totalRecords'], 0);
+
+    if (!stripes.hasPerm('ui-users.requests.all,ui-requests.all')) return openRequestsCount;
+
+    const openRequestStatuses = [
+      'Open - Not yet filled',
+      'Open - Awaiting pickup',
+      'Open - In transit'
+    ]
+      .map(status => `requestStatus.${status}`)
+      .join('&');
+
+    const openRequestsPath = `/requests?query=${user.barcode}&filters=${openRequestStatuses}&sort=Request date`;
+
+    return (
+      <Link
+        data-test-open-requests-count
+        to={openRequestsPath}
+      >
+        {openRequestsCount}
+      </Link>
     );
   }
 
@@ -125,7 +177,7 @@ class UserDetail extends React.Component {
     const patronGroups = (resources.patronGroups || {}).records || [];
     const patronGroup = patronGroups[0] || {};
     const hasProfilePicture = !!(settings.length && settings[0].value === 'true');
-    const statusVal = (_.get(user, ['active'], '') ? 'ui-checkout.active' : 'ui-checkout.inactive');
+    const statusVal = (get(user, ['active'], '') ? 'ui-checkout.active' : 'ui-checkout.inactive');
 
     return (
       <div>